v3.0: No Parent/Child queries, "invisibility map", linked index improvements, and PG 9.[345] support (#120)
* A squashmerge of both the "v2.7.0" branch *and* the "pg95-support" branch along with changes to support PG 9.3, 9.4, and 9.5 all in one codebase.

v2.7 is going to die and we're going to move forward with v3.0!

* fix travis

* cleanup

* doc cleanup

* (finally) document shadow indexes

* doc updates

* doc fix

* fix type-o

* docs update

* docs update

* docs update

* docs update

* docs update

* docs update

* code cleanup

* fix type-o

* reformat code

* reformat code

* update script

* docs update

* docs update

* docs update

* docs update

+++ postgres/src/main/shell/build-packages.sh
@@ -3,39 +3,42 @@
 VERSION=$1
 BASE=`pwd`
 DISTROS="centos6 centos7 ubuntu_trusty ubuntu_precise"
-POSTGRES_VERSION="9.3"
+POSTGRES_VERSIONS="9.3 9.4 9.5"
 
 ##
 # compile ZomboDB for target distros
 ##
-for distro in ${DISTROS} ; do
-    cd $BASE
+for POSTGRES_VERSION in ${POSTGRES_VERSIONS} ; do
+    for distro in ${DISTROS} ; do
+        cd $BASE
 
-    mkdir -p $BASE/target/${distro}
+        mkdir -p $BASE/target/pg${POSTGRES_VERSION}/${distro}
 
-    cd src/main/docker/zombodb-build-${distro}
+        cd src/main/docker/pg${POSTGRES_VERSION}/zombodb-build-${distro}
 
-    docker build -t zombodb-build-${POSTGRES_VERSION}-${distro} .
-    docker run --rm -v $BASE:/mnt -w /mnt -e DESTDIR=target/${distro} zombodb-build-${POSTGRES_VERSION}-${distro} make clean install
+        docker build -t zombodb-build-${POSTGRES_VERSION}-${distro} .
+        docker run --rm -v $BASE:/mnt -w /mnt -e DESTDIR=target/pg${POSTGRES_VERSION}/${distro} zombodb-build-${POSTGRES_VERSION}-${distro} make clean install
 
-    # move the zombod.so into the plugins/ directory
-    cd $BASE/target
-    cd `dirname $(find ${distro} -name "zombodb.so")`
-    mkdir plugins
-    cd plugins
-    mv ../zombodb.so .
+        # move the zombod.so into the plugins/ directory
+        cd $BASE/target/pg${POSTGRES_VERSION}
+        cd `dirname $(find ${distro} -name "zombodb.so")`
+        mkdir plugins
+        cd plugins
+        mv ../zombodb.so .
+
+        ##
+        # also build a tarball from the Ubuntu_precise version
+        ##
+        if [ $distro == "ubuntu_precise" ] ; then
+            cd $BASE
+            cd target/pg${POSTGRES_VERSION}
+            rm -rf tarball
+            mkdir -p tarball/lib tarball/share
+            cp -Rp ubuntu_precise/usr/lib/postgresql/${POSTGRES_VERSION}/lib/* tarball/lib
+            cp -Rp ubuntu_precise/usr/share/postgresql/${POSTGRES_VERSION}/* tarball/share
+            cd tarball/
+            tar czf ../zombodb-precise-pg${POSTGRES_VERSION}-${VERSION}.tgz .
+        fi
+    done
+
 done
-
-
-##
-# also build a tarball from the Ubuntu_precise version
-##
-
-cd $BASE
-cd target
-rm -rf tarball
-mkdir -p tarball/lib tarball/share
-cp -Rp ubuntu_precise/usr/lib/postgresql/${POSTGRES_VERSION}/lib/* tarball/lib
-cp -Rp ubuntu_precise/usr/share/postgresql/${POSTGRES_VERSION}/* tarball/share
-cd tarball/
-tar czf ../zombodb-precise-pg93-$VERSION.tgz .
